engine/device_manager.py
class DeviceManager():
    devices: dict = {}
    devices_status: dict = {}
    device_id: dict = {}

    def __init__(self):
        pass

    def open_all_devices(self) -> bool:
        return True

    def start_all_devices(self) -> bool:
        return True

    def stop_all_devices(self) -> bool:
        return True

    def close_all_devices(self) -> bool:
        return True

    def load_devices_from_db(self) -> bool:
        return True

    def get_all_devices_status(self) -> list:
        return [True]

    def get_device_status(self, dev_name: str) -> bool:
        return True

    def update_device_property(self, device: dict) -> bool:
        return True

    def is_device_open(self, dev_name: str) -> bool:
        return True

    def test_all_devices(self) -> list:
        return [{"device_name": "xyz", "status": True}]

    def test_device(self, dev: dict) -> bool:
        return True

    def access_device(self, dev: dict) -> bool:
        return True

    def get_failed_devices(self) -> list:
        return [True]
